+++ drv_imu_inv_icm20602.h
@@ -18,7 +18,7 @@ typedef struct __inv_icm20602 {
 
 
 inline void ICM20602_Destruct(inv_icm20602_handle _this) { _IMU_Destruct((void *) _this); }
-inv_icm20602_handle ICM20602_Construct(inv_i2c _i2c, uint16_t _addr);
+inv_icm20602_handle ICM20602_Construct(inv_i2c _i2c, uint8_t _addr);
 inv_icm20602_handle ICM20602_Construct2(inv_spi _spi);
