+++ drv_imu_inv_mpu6050.c
@@ -37,7 +37,7 @@ inv_imu_vector_table mpu6050_VectorTable = {
         .Destruct = (void*) MPU6050_Destruct
 };
 
-inv_mpu6050_handle MPU6050_Construct(inv_i2c _i2c, uint16_t _addr) {
+inv_mpu6050_handle MPU6050_Construct(inv_i2c _i2c, uint8_t _addr) {
     inv_mpu6050_handle rtv = (void *) INV_REALLOC(IMU_Construct(_i2c, _addr), sizeof(inv_mpu6050));
     memset((void *) ((char *)rtv + sizeof(inv_mpu6050) - sizeof(inv_imu)), 0, sizeof(inv_mpu6050) - sizeof(inv_imu));
     rtv->parents.vtable = &mpu6050_VectorTable;
